int to long

+++ src/main/java/com/example/shawarmahouse/model/AllOrders.java
@@ -8,7 +8,6 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.HashMap;
 
 @Document(collection = "allorders")
@@ -32,7 +31,7 @@ public class AllOrders {
 
     private HashMap<String, Integer> itemsWithQuantity;
 
-    private int totalAmount;
+    private long totalAmount;
 
 
 }
